+++ frontend/src/components/Calendar/Calendar.js
@@ -171,7 +171,7 @@ class Calendar extends Component {
                             day: "Jour",
                             date: "Date",
                             time: "Heure",
-                            event: "Événement",
+                            event: "Évènement",
                         }}
                         eventPropGetter={eventStyleGetter}
                         onSelectEvent={this.openEventModal}