+++ frontend/src/components/ClientForm/ClientForm.js
@@ -153,7 +153,7 @@ class ClientForm extends Component {
                 >
                     <div className={styles.selectedCategory}>
                         <label className={styles.labelRadio}>
-                            Type de client:
+                            Type de client :
                         </label>
                         <div>
                             <label className={styles.radioParticulier}>