+++ frontend/src/pages/ClientDetails/ClientDetailsPage.js
@@ -5,7 +5,6 @@ import TemplateGlobal from "../Template/TemplateGlobal";
 import TemplateGlobalMobile from "../Template/TemplateGlobalMobile";
 import styles from "./ClientDetailsPage.module.css";
 import stylesMobile from "./ClientDetailsPageMobile.module.css";
-import profilPicture from "../../images/profil.png";
 import { getClientById } from "../../services/api";
 
 function ClientDetailsPageWrapper() {
@@ -269,10 +268,9 @@ class ClientDetailsPage extends Component {
                                         <div
                                             className={styles.profilCompanyBack}
                                         >
-                                            <img
-                                                src={profilPicture}
-                                                alt="Profil picture"
-                                            />
+                                            <div className={styles.iconWrapper}>
+                                                <i className="fa-solid fa-landmark"></i>
+                                            </div>
                                             <p className={styles.company}>
                                                 {client.company}
                                             </p>
@@ -370,10 +368,9 @@ class ClientDetailsPage extends Component {
                                             Détails client
                                         </h1>
                                         <div className={styles.profilNamesBack}>
-                                            <img
-                                                src={profilPicture}
-                                                alt="Profil picture"
-                                            />
+                                            <div className={styles.iconWrapper}>
+                                                <i className="fa-solid fa-user"></i>
+                                            </div>
                                             <div className={styles.nameColumn}>
                                                 <p className={styles.lastname}>
                                                     {client.lastname}{" "}