+++ frontend/src/pages/EmployeeDetails/EmployeeDetailsPage.js
@@ -5,7 +5,6 @@ import TemplateGlobal from "../Template/TemplateGlobal";
 import TemplateGlobalMobile from "../Template/TemplateGlobalMobile";
 import styles from "./EmployeeDetailsPage.module.css";
 import stylesMobile from "./EmployeeDetailsPageMobile.module.css";
-import profilPicture from "../../images/profil.png";
 import { getEmployeeById } from "../../services/api";
 
 function EmployeeDetailsPageWrapper() {
@@ -112,10 +111,9 @@ class EmployeeDetailsPage extends Component {
                             <div className={styles.profilInfo}>
                                 <h1 className={styles.pageTitle}>Employé</h1>
                                 <div className={styles.alignBackButton}>
-                                    <img
-                                        src={profilPicture}
-                                        alt="Profil picture"
-                                    />
+                                    <div className={styles.iconWrapper}>
+                                        <i className="fa-solid fa-user"></i>
+                                    </div>
                                     <div className={styles.names}>
                                         <p className={styles.firstname}>
                                             {employee.firstname}{" "}