+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 -   Fix event unit tests
 -   Fix employee unit tests
+-   Add Appointment and Intervention classes unit tests
 
 # [0.8.0] (2025-21-02)