+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changes
 
 -   Revised .env variables to improve the deployment process
+-   UI updates
 
 ## Fixes